% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archivist-package.R
\docType{package}
\name{archivist}
\alias{archivist}
\alias{archivist-package}
\title{\code{archivist} package}
\description{
Extract information from R package archives.
}
\details{
See the README on \href{https://github.com/robertdj/archivist}{GitHub}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertdj/archivist}
}

}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{git@dahl-jacobsen.dk}

}
