% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{get_desc_field}
\alias{get_desc_field}
\title{Get fields from DESCRIPTION file}
\usage{
get_desc_field(archive, field)
}
\arguments{
\item{archive}{\verb{[character]} The path of the archive.}

\item{field}{\verb{[character]} A vector of field names to return.}
}
\value{
The fields in a named vector.
}
\description{
Get fields from DESCRIPTION file
}
