% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_file_in_archive}
\alias{get_file_in_archive}
\title{Get file in archive}
\usage{
get_file_in_archive(package_file, archive)
}
\arguments{
\item{package_file}{\verb{[character]} The file inside the archive.}

\item{archive}{\verb{[character]} The path of the archive.}
}
\value{
The path to the \emph{extracted} \code{package_file}.
}
\description{
Get a file in a compressed archive.
}
\details{
The extracted \code{package_file} is in the R session's temporary folder and is \emph{not} deleted
automatically (before the temporary folder is deleted upon ending the R session). Cleaning up is
left to the caller.
}
