% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{get_package_desc}
\alias{get_package_desc}
\title{Get package DESCRIPTION file}
\usage{
get_package_desc(archive)
}
\arguments{
\item{archive}{\verb{[character]} The path of the archive.}
}
\value{
The package's DESCRIPTION file parsed by \code{\link[=read.dcf]{read.dcf()}}
}
\description{
Get package DESCRIPTION file
}
