% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name.R
\name{get_package_name}
\alias{get_package_name}
\title{Get name of package}
\usage{
get_package_name(archive)
}
\arguments{
\item{archive}{\verb{[character]} The path of the archive.}
}
\value{
A character.
}
\description{
Get the name of the package as it is specified in the DESCRIPTION file.
}
