% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{get_package_version}
\alias{get_package_version}
\title{Get version of package}
\usage{
get_package_version(archive)
}
\arguments{
\item{archive}{\verb{[character]} The path of the archive.}
}
\value{
A \code{\link[=package_version]{package_version()}}.
}
\description{
Get the version of the package as it is specified in the DESCRIPTION file.
}
