% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\name{get_r_version}
\alias{get_r_version}
\title{Get R version used for build}
\usage{
get_r_version(archive)
}
\arguments{
\item{archive}{\verb{[character]} The path of the archive.}
}
\value{
The package as a \code{R_system_version} -- see \code{\link[=getRversion]{getRversion()}}.
}
\description{
Get the R version used to build the package archive.
}
