% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desc.R
\name{is_package_built}
\alias{is_package_built}
\title{Is package built?}
\usage{
is_package_built(archive)
}
\arguments{
\item{archive}{\verb{[character]} The path of the archive.}
}
\value{
A boolean.
}
\description{
Determine if a package archive is built/compiled. Only relevant for \code{tar.gz} archives (on Linux).
}
