% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{package_ext}
\alias{package_ext}
\title{Get file extension of package archive}
\usage{
package_ext(archive_file)
}
\arguments{
\item{archive_file}{Path to a package archive.}
}
\value{
The file extension.
}
\description{
Get the file extension of a package archive if it is valid (\code{tar.gz}, \code{tgz} or \code{zip}).
An error is thrown if the extension is not valid.
}
