% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgpeek-package.R
\docType{package}
\name{pkg.peek}
\alias{pkg.peek}
\alias{pkg.peek-package}
\title{\code{pkg.peek} package}
\description{
Extract information from R package archives.
}
\details{
See the README on \href{https://github.com/robertdj/pkg.peek}{GitHub}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/robertdj/pkg.peek}
}

}
\author{
\strong{Maintainer}: Robert Dahl Jacobsen \email{git@dahl-jacobsen.dk}

}
