% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{get_file_in_archive}
\alias{get_file_in_archive}
\title{Get file in archive}
\usage{
get_file_in_archive(archive, package_file)
}
\arguments{
\item{archive}{\verb{[character]} The path of the archive.}

\item{package_file}{\verb{[character]} The file inside the archive.}
}
\value{
The path to the \emph{extracted} \code{package_file}. This will be in the R session's temporary
folder.
}
\description{
Get a file in a compressed archive.
}
